Implement window server and mouse/keyboard event queue

+++ src/include/ctype.h
@@ -14,10 +14,10 @@ typedef unsigned long long ino_t;
 typedef unsigned short mode_t;
 typedef long ssize_t;
 typedef unsigned long sector_t;
-typedef unsigned short pid_t;
-typedef unsigned short tid_t;
-typedef unsigned short uid_t;
-typedef unsigned short gid_t;
+typedef unsigned int pid_t;
+typedef unsigned int tid_t;
+typedef unsigned int uid_t;
+typedef unsigned int gid_t;
 
 #define isspace(c) ((c) == ' ' || ((c) >= '\t' && (c) <= '\r'))
 #define isupper(c) ((c) >= 'A' && (c) <= 'Z')
